move visual elements to core, remove fluent drawing
