+++ src/skiasharp/LiveChartsCore.SkiaSharp/VisualElements/GeometryVisual.cs
@@ -20,9 +20,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-using System;
 using LiveChartsCore.Drawing;
-using LiveChartsCore.Measure;
 using LiveChartsCore.SkiaSharpView.Drawing;
 using LiveChartsCore.VisualElements;
 
@@ -31,55 +29,6 @@ namespace LiveChartsCore.SkiaSharpView.VisualElements;
 /// <summary>
 /// Defines a visual element in a chart that draws a sized geometry in the user interface.
 /// </summary>
-public class GeometryVisual<TGeometry> : BaseGeometryVisual
+public class GeometryVisual<TGeometry> : GeometryVisual<TGeometry, SkiaSharpDrawingContext>
     where TGeometry : ISizedGeometry<SkiaSharpDrawingContext>, new()
-{
-    internal readonly TGeometry _geometry = new();
-
-    internal override IAnimatable?[] GetDrawnGeometries()
-    {
-        return new IAnimatable?[] { _geometry };
-    }
-
-    /// <inheritdoc cref="VisualElement{TDrawingContext}.OnInvalidated(Chart{TDrawingContext})"/>
-    protected internal override void OnInvalidated(Chart<SkiaSharpDrawingContext> chart)
-    {
-        var l = GetActualCoordinate();
-
-        var size = Measure(chart);
-
-        _geometry.X = l.X;
-        _geometry.Y = l.Y;
-        _geometry.Width = size.Width;
-        _geometry.Height = size.Height;
-
-        var drawing = chart.Canvas.Draw();
-        if (Fill is not null) _ = drawing.SelectPaint(Fill).Draw(_geometry);
-        if (Stroke is not null) _ = drawing.SelectPaint(Stroke).Draw(_geometry);
-    }
-
-    /// <inheritdoc cref="VisualElement{TDrawingContext}.SetParent(IGeometry{TDrawingContext})"/>
-    protected internal override void SetParent(IGeometry<SkiaSharpDrawingContext> parent)
-    {
-        if (_geometry is null) return;
-        _geometry.Parent = parent;
-    }
-
-    /// <inheritdoc cref="VisualElement{TDrawingContext}.Measure(Chart{TDrawingContext})"/>
-    public override LvcSize Measure(Chart<SkiaSharpDrawingContext> chart)
-    {
-        var w = (float)Width;
-        var h = (float)Height;
-
-        if (SizeUnit == MeasureUnit.ChartValues)
-        {
-            if (PrimaryScaler is null || SecondaryScaler is null)
-                throw new Exception($"You can not use {MeasureUnit.ChartValues} scale at this element.");
-
-            w = SecondaryScaler.MeasureInPixels(w);
-            h = PrimaryScaler.MeasureInPixels(h);
-        }
-
-        return new LvcSize(w, h);
-    }
-}
+{ }
