+++ src/skiasharp/LiveChartsCore.SkiaSharp/VisualElements/LabelVisual.cs
@@ -20,157 +20,12 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-using System;
-using LiveChartsCore.Drawing;
-using LiveChartsCore.Measure;
 using LiveChartsCore.SkiaSharpView.Drawing;
 using LiveChartsCore.SkiaSharpView.Drawing.Geometries;
 using LiveChartsCore.VisualElements;
 
 namespace LiveChartsCore.SkiaSharpView.VisualElements;
 
-/// <summary>
-/// Defines a visual element with stroke and fill properties.
-/// </summary>
-public class LabelVisual : VisualElement<SkiaSharpDrawingContext>
-{
-    internal readonly LabelGeometry _labelGeometry = new();
-    internal IPaint<SkiaSharpDrawingContext>? _paint;
-    internal bool _isVirtual = false;
-    internal string _text = string.Empty;
-    internal double _textSize = 12;
-    internal Align _verticalAlignment = Align.Middle;
-    internal Align _horizontalAlignment = Align.Middle;
-    internal LvcColor _backgroundColor;
-    internal Padding _padding = new(0);
-    internal double _rotation;
-    internal float _lineHeight = 1.75f;
-    internal LvcPoint _translate = new();
-
-    /// <summary>
-    /// Gets or sets the fill paint.
-    /// </summary>
-    public IPaint<SkiaSharpDrawingContext>? Paint
-    {
-        get => _paint;
-        set => SetPaintProperty(ref _paint, value);
-    }
-
-    /// <summary>
-    /// Gets or sets the label text.
-    /// </summary>
-    public string Text { get => _text; set => SetProperty(ref _text, value); }
-
-    /// <summary>
-    /// Gets or sets the font size.
-    /// </summary>
-    public double TextSize { get => _textSize; set => SetProperty(ref _textSize, value); }
-
-    /// <summary>
-    /// Gets or sets the rotation.
-    /// </summary>
-    public double Rotation { get => _rotation; set { _rotation = value; OnPropertyChanged(); } }
-
-    /// <summary>
-    /// Gets or sets the translate transform.
-    /// </summary>
-    public LvcPoint Translate { get => _translate; set => SetProperty(ref _translate, value); }
-
-    /// <summary>
-    /// Gets or sets the vertical alignment.
-    /// </summary>
-    public Align VerticalAlignment { get => _verticalAlignment; set => SetProperty(ref _verticalAlignment, value); }
-
-    /// <summary>
-    /// Gets or sets the horizontal alignment.
-    /// </summary>
-    public Align HorizontalAlignment { get => _horizontalAlignment; set => SetProperty(ref _horizontalAlignment, value); }
-
-    /// <summary>
-    /// Gets or sets the background color.
-    /// </summary>
-    public LvcColor BackgroundColor { get => _backgroundColor; set => SetProperty(ref _backgroundColor, value); }
-
-    /// <summary>
-    /// Gets or sets the padding.
-    /// </summary>
-    public Padding Padding { get => _padding; set => SetProperty(ref _padding, value); }
-
-    /// <summary>
-    /// Gets or sets the line height [times the text measured height].
-    /// </summary>
-    public float LineHeight { get => _lineHeight; set => SetProperty(ref _lineHeight, value); }
-
-    internal override IPaint<SkiaSharpDrawingContext>?[] GetPaintTasks()
-    {
-        return new[] { _paint };
-    }
-
-    internal override IAnimatable?[] GetDrawnGeometries()
-    {
-        return new IAnimatable?[] { _labelGeometry };
-    }
-
-    internal override void AlignToTopLeftCorner()
-    {
-        VerticalAlignment = Align.Start;
-        HorizontalAlignment = Align.Start;
-    }
-
-    /// <inheritdoc cref="VisualElement{TDrawingContext}.OnInvalidated(Chart{TDrawingContext})"/>
-    protected internal override void OnInvalidated(Chart<SkiaSharpDrawingContext> chart)
-    {
-        var x = (float)X;
-        var y = (float)Y;
-
-        if (LocationUnit == MeasureUnit.ChartValues)
-        {
-            if (PrimaryScaler is null || SecondaryScaler is null)
-                throw new Exception($"You can not use {MeasureUnit.ChartValues} scale at this element.");
-
-            x = SecondaryScaler.ToPixels(x);
-            y = PrimaryScaler.ToPixels(y);
-        }
-
-        _ = Measure(chart);
-
-        _labelGeometry.Text = Text;
-        _labelGeometry.TextSize = (float)TextSize;
-        _labelGeometry.X = x;
-        _labelGeometry.Y = y;
-        _labelGeometry.RotateTransform = (float)Rotation;
-        _labelGeometry.TranslateTransform = Translate;
-        _labelGeometry.VerticalAlign = VerticalAlignment;
-        _labelGeometry.HorizontalAlign = HorizontalAlignment;
-        _labelGeometry.Background = BackgroundColor;
-        _labelGeometry.Padding = Padding;
-        _labelGeometry.LineHeight = LineHeight;
-
-        var drawing = chart.Canvas.Draw();
-        if (Paint is not null) _ = drawing.SelectPaint(Paint).Draw(_labelGeometry);
-    }
-
-    /// <inheritdoc cref="VisualElement{TDrawingContext}.SetParent(IGeometry{TDrawingContext})"/>
-    protected internal override void SetParent(IGeometry<SkiaSharpDrawingContext> parent)
-    {
-        if (_labelGeometry is null) return;
-        _labelGeometry.Parent = parent;
-    }
-
-    /// <inheritdoc cref="VisualElement{TDrawingContext}.Measure(Chart{TDrawingContext})"/>
-    public override LvcSize Measure(Chart<SkiaSharpDrawingContext> chart)
-    {
-        _labelGeometry.Text = Text;
-        _labelGeometry.TextSize = (float)TextSize;
-        _labelGeometry.RotateTransform = (float)Rotation;
-        _labelGeometry.TranslateTransform = Translate;
-        _labelGeometry.VerticalAlign = VerticalAlignment;
-        _labelGeometry.HorizontalAlign = HorizontalAlignment;
-        _labelGeometry.Background = BackgroundColor;
-        _labelGeometry.Padding = Padding;
-
-        return _paint is null
-            ? new LvcSize()
-            : _labelGeometry.Measure(_paint);
-    }
-}
+/// <inheritdoc cref="LabelVisual   {TLabelGeometry, TDrawingContext}"/>
+public class LabelVisual : LabelVisual<LabelGeometry, SkiaSharpDrawingContext>
+{ }
